still far

+++ tests/cdogrid_test.py
@@ -12,21 +12,21 @@ cdo = Cdo()
     ("r360x180", True),          # Global regular NxM grid
     ("lon=-75.0/lat=40.0", True),  # One grid point
     ("F64", True),               # Full regular Gaussian grid
-    ("N400", True),              # Gaussian reduced grid
+    ("N200", True),              # Gaussian reduced grid
     ("f64", True),               # Full regular Gaussian grid
-    ("n400", True),              # Gaussian reduced grid
+    ("n200", True),              # Gaussian reduced grid
     ("gme10", True),             # Global icosahedral-hexagonal GME grid
-    ("hp1024", True),            # HEALPix grid
+    ("hp256", True),            # HEALPix grid
     ("hp32_ring", True),         # HEALPix grid
     ("hpz4", True),              # HEALPix zoom grid
-    ("hpr16", True),             # HEALPix refinement grid
+    ("hpr6", True),             # HEALPix refinement grid
     ("random_string", False),    # Invalid CDO grid
     ("/path/to/file.nc", False)  # Invalid CDO grid (file path)
 ])
 def test_is_cdo_grid_and_recognition(grid_str, expected, tmp_path):
     """Simple test for cdo grid detection"""
     assert bool(CdoGrid(grid_str).grid_kind) == expected
-    if expected:
+    if expected and 'dcw' not in grid_str:
         cdo.const(f'1,{grid_str}', output=str(tmp_path / "output.nc"), loglevel='debug')
         assert os.path.exists(tmp_path / "output.nc")
         os.remove(tmp_path / "output.nc")
